HT_24/mysite/requirements.txt
celery==5.3.6
Django==5.0.1
factory-boy==3.3.0
djangorestframework==3.14.0
redis==5.0.1
requests==2.31.0
